examples/python/openai-agents-sdk/requirements.txt
openai-agents>=0.0.3
python-dotenv>=1.0.0

traceroot==0.1.0
openinference-instrumentation-openai-agents>=0.1.0
opentelemetry-sdk>=1.0.0
